refactor: remove command_register and add it as model methods
